Document dense examples
