+++ examples/README.md
@@ -1,0 +1,9 @@
+# Examples
+
+This directory provides some simple examples using linalgcpp
+
+Current Examples:
+* ['Vector'](vector)
+* ['DenseMatrix'](dense)
+* ['SparseMatrix'](sparse)
+
